Add domain verification
